Added type cast to avoid compiler warning on MSVC5.

+++ dcmimgle/libsrc/dibaslut.cc
@@ -22,9 +22,9 @@
  *  Purpose: DicomBaseLUT (Source)
  *
  *  Last Update:      $Author: joergr $
- *  Update Date:      $Date: 1999-03-03 12:05:43 $
+ *  Update Date:      $Date: 1999-03-03 17:56:12 $
  *  Source File:      $Source: /export/gitmirror/dcmtk-git/../dcmtk-cvs/dcmtk/dcmimgle/libsrc/dibaslut.cc,v $
- *  CVS/RCS Revision: $Revision: 1.2 $
+ *  CVS/RCS Revision: $Revision: 1.3 $
  *  Status:           $State: Exp $
  *
  *  CVS/RCS Log at end of file
@@ -76,7 +76,7 @@ int DiBaseLUT::invertTable()
         register Uint32 i;
         register const Uint16 *p = Data;
         register Uint16 *q = (Uint16 *)Data;                // remove const to modify Data
-        const Uint16 max = DicomImageClass::maxval(Bits);
+        const Uint16 max = (Uint16)DicomImageClass::maxval(Bits);
         for (i = 0; i < Count; i++)
             *(q++) = max - *(p++);
         return 1;
@@ -89,7 +89,10 @@ int DiBaseLUT::invertTable()
  *
  * CVS/RCS Log:
  * $Log: dibaslut.cc,v $
- * Revision 1.2  1999-03-03 12:05:43  joergr
+ * Revision 1.3  1999-03-03 17:56:12  joergr
+ * Added type cast to avoid compiler warning on MSVC5.
+ *
+ * Revision 1.2  1999/03/03 12:05:43  joergr
  * Added method to invert lookup table data (used for presentation state LUTs).
  *
  * Revision 1.1  1999/02/03 17:47:44  joergr
